pyproject.toml
[project]
name = "TODO-project"
version = "0.1.0"
description = "study-project"
authors = [
    {name = "Rabbau",email = "oprovozenko@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
